+++ assets/js/components/button-page/index.js
@@ -1,3 +1,3 @@
-import ButtonPage from "./button-page-hooks";
+import ButtonPage from "./button-page";
 
 export default ButtonPage;
