--- assets/js/modules/gate-button/gate-button.less
@@ -1,52 +1,0 @@
-.round-button {
-  & ._note {
-    font-size: 150px;
-    width: 1em;
-    height: 1em;
-    border-radius: 50%;
-    background-color: #f5f5f5;
-    color: rgb(150, 150, 150);
-    text-align: center;
-    overflow: hidden;
-    z-index: 4;
-    transition: transform 0.2s ease-in-out, box-shadow 0.2s linear;
-    box-shadow: 0 0 0.2em rgba(204, 204, 204, 0.7);
-
-    & span {
-      font-size: 0.2em;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      height: 100%;
-      user-select: none;
-      transition: transform 0.2s ease-in-out;
-    }
-  }
-
-  &.is-active {
-    &:hover {
-      cursor: default;
-
-      & ._note {
-        box-shadow: 0 0 0.15em 2px rgba(225, 225, 225, 0.4);
-
-        & > span {
-          color: rgb(100, 100, 100);
-        }
-      }
-    }
-
-    &:active {
-      cursor: default;
-
-      & ._note {
-        box-shadow: 0 0 0.4em 4px rgb(225, 225, 225);
-        transform: scale(0.9);
-
-        & > span {
-          transform: scale(0.95);
-        }
-      }
-    }
-  }
-}
