--- assets/js/modules/messager/message.less
@@ -1,104 +1,0 @@
-.popup {
-  &-container {
-    margin: 0 auto;
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: auto;
-    z-index: 100;
-  }
-
-  &-message {
-    display: flex;
-    align-items: center;
-    margin: 0;
-    padding: 0;
-    font-size: 1rem;
-    font-weight: 500;
-    box-shadow: 0 1rem 10px rgba(0,0,0,.4);
-    background-color: white;
-    border-radius: 0.4rem;
-    overflow: hidden;
-    color: white;
-    opacity: 0;
-    transition: all 0.3s ease-in-out;
-    position: relative;
-    left: 0;
-    top: -2rem;
-    visibility: hidden;
-    height: 0;
-
-
-    &:before {
-      font-family: "Font Awesome 5 Free", sans-serif, "Font Awesome 5 Brands", sans-serif;
-      font-size: 1.2rem;
-      content: '';
-      display: inline-block;
-      text-align: center;
-      height: 2.2rem;
-      width: 2.2rem;
-      min-height: 2.2rem;
-      min-width: 2.2rem;
-      line-height: 2rem;
-      background-color: white;
-      border: 2px solid white;
-      border-radius: 50%;
-      margin: 0 .5rem 0 1.5rem;
-      background-color: rgba(255, 255, 255, 0.3);
-    }
-
-    &.is-show {
-      opacity: 0.9;
-      visibility: visible;
-      top: 0;
-      height: 3.2rem;
-      margin: 3px auto;
-    }
-
-    &.success {
-      background: rgba(0,142,0,.90);
-      border-color: #004C00;
-
-      &:before {
-        content: ''; //@fa-var-check
-      }
-    }
-
-    &.error {
-      background: rgba(186,0,0,.90);
-      border-color: #500000;
-
-      &:before {
-        content: ''; //@fa-var-times
-      }
-    }
-
-    &.info {
-      background: rgba(12,95,178,.90);
-      border-color: #05284C;
-
-      &:before {
-        content: ''; // @fa-var-info
-      }
-    }
-
-    &.warning {
-      background: rgba(255, 151, 2, .90);
-      border-color: #573401;
-
-      &:before {
-        content: ''; //@fa-var-exclamation-triangle
-      }
-    }
-
-    &.debug {
-      background: black;
-      border-color: black;
-
-      &:before {
-        content: ''; //@fa-var-bug
-      }
-    }
-  }
-}
