--- assets/js/modules/progress-bar/progress-button.less
@@ -1,61 +1,0 @@
-.progress-bar {
-  font-size: 180px;
-  width: 1em;
-  height: 1em;
-  border-radius: 50%;
-  background-color: rgb(204, 204, 204);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  overflow: hidden;
-  position: relative;
-  box-shadow: 0 0 0.1em rgba(255, 0, 0, 0.7);
-
-  & ._line {
-    position: absolute;
-    top: 0;
-    left: 0;
-    height: 100%;
-    width: 50%;
-    background-color: rgb(204, 204, 204);
-    transform-origin: 100% 50%;
-    transform: rotate(0deg);
-
-  }
-
-  & ._r-mask {
-    visibility: hidden;
-    position: absolute;
-    top: 0;
-    left: 50%;
-    width: 50%;
-    height: 100%;
-    background-color: #5eb5e0;
-
-  }
-
-  &:before {
-    visibility: visible;
-    content: '';
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 50%;
-    height: 100%;
-    background-color: inherit;
-    z-index: 2;
-  }
-
-}
-
-.gt50 {
-  &.progress-bar {
-    &:before {
-      visibility: hidden;
-    }
-
-  }
-  & ._r-mask {
-    visibility: visible;
-  }
-}
