--- assets/js/modules/sharing-button/sharing-button.less
@@ -1,198 +1,0 @@
-.sharing {
-
-  &-wrapper {
-    display: none;
-    bottom: 1.2rem;
-    opacity: 0;
-    width: 100%;
-    height: 100%;
-    background: rgb(0,0,0);
-    background: linear-gradient(0deg, rgba(0,0,0,0.8) 70%, rgba(0,0,0,0) 100%);
-    position: absolute;
-    left: 0;
-    padding: 150px 0 100px;
-    transition: all 0.3s ease-in-out;
-
-    &.is-inited {
-      display: block;
-    }
-
-    &.is-active {
-      bottom: 0;
-      opacity: 1;
-    }
-
-  }
-
-  &-list {
-    width: 100%;
-    height: 100%;
-    padding: 0;
-    margin: 0 auto;
-    list-style: none;
-    display: flex;
-    flex-wrap: nowrap;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    z-index: 10;
-  }
-
-  &-item {
-    font-weight: 500;
-    position: relative;
-    display: block;
-    height: 2.5rem;
-    line-height: 2.5rem;
-    width: 100%;
-    min-width: 200px;
-    max-width: 300px;
-    background-color: #F5F5F5;
-    border: 1px solid grey;
-    border-radius: 10px;
-    margin-bottom: 1rem;
-    box-shadow: 2px -2px 5px 0 rgba(0,0,0,.1),
-                -2px -2px 5px 0 rgba(0,0,0,.1),
-    2px 2px 5px 0 rgba(0,0,0,.1),
-                -2px 2px 5px 0 rgba(0,0,0,.1);
-    border-right: 10px solid rgb(3, 146, 85);
-    border-left: 10px solid rgb(3, 146, 85);
-    transition: all 0.3s ease-in-out;
-
-    &:last-child  {
-      border-left: 10px solid rgb(194, 4, 55);
-      border-right: 10px solid rgb(194, 4, 55);
-      background: rgb(238, 227, 227);
-    }
-
-    &:active {
-      transform: translate3d(0, 3px, 3px) scale(0.97);
-    }
-
-    &:hover {
-      transform: translate3d(0, -3px, 3px) scale(1.03);
-    }
-
-    & > a {
-      display: block;
-      height: 100%;
-      width: 100%;
-      padding-left: 1.5rem;
-      text-decoration: none;
-      color: rgb(100, 100, 100);
-      -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
-    }
-
-  }
-
-  &-button {
-    margin: 0 auto;
-    padding: 1rem 0;
-    display: flex;
-    flex-direction: column;
-    flex-wrap: nowrap;
-    justify-content: flex-end;
-    align-items: center;
-
-    &__btn {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      margin: 0 auto;
-      min-width: 4rem;
-      min-height: 4rem;
-      max-width: 4rem;
-      max-height: 4rem;
-      border-radius: 50%;
-      border: 3px solid #a3a3a3;
-      background: #f5f5f5;
-      text-decoration: none;
-      color: rgb(150, 150, 150);
-      font-size: 2rem;
-      box-shadow: #a3a3a3 0 0 10px;
-      transition: 0.3s ease-in-out;
-      -webkit-tap-highlight-color: rgba(0, 0, 0, 0);
-
-      &:before {
-        font-family: "Font Awesome 5 Free", sans-serif;
-        content: ''; //@fa-var-copy
-        position: relative;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100%;
-      }
-
-      &:active {
-        color: #646464;
-        border-color: #F5F5F5;
-        box-shadow: #a3a3a3 0 0 20px;
-      }
-
-      &:hover {
-        color: #646464;
-        border-color: #F5F5F5;
-        box-shadow: #a3a3a3 0 0 20px;
-      }
-    }
-
-    &__label {
-      visibility: hidden;
-      color: #f5f5f5;
-      font-size: 1.2rem;
-
-    }
-
-    &__input {
-      border: 1px solid #646464;
-      padding: 5px 10px;
-      visibility: hidden;
-      border-radius: 10px;
-      height: 2rem;
-      margin-bottom: 10px;
-      width: 100%;
-      max-width: 450px;
-      color: #646464;
-
-      &:focus {
-        outline: none;
-      }
-
-    }
-
-    &.share {
-      & .sharing-button {
-
-        &__label {
-          visibility: visible;
-        }
-
-        &__input {
-          visibility: visible;
-        }
-
-        &__btn {
-
-          &:before {
-            content: ''; //@fa-var-copy
-          }
-        }
-      }
-    }
-  }
-}
-
-
-@media (max-width: 960px) {
-  .sharing {
-
-    &-wrapper {
-      padding: 50px 0 80px;
-    }
-
-    &-list {
-      justify-content: flex-end;
-    }
-
-  }
-}
