+++ assets/less/main.less
@@ -7,5 +7,3 @@
 @import "sections/sec_login";
 
 @import "sections/sec_button";
-
-@import "sections/sec_access_denied";
