--- assets/less/sections/sec_access_denied.less
@@ -1,9 +1,0 @@
-.sec-access-denied {
-
-  &__text {
-    font-weight: 500;
-    font-size: 1rem;
-    align-content: center;
-  }
-}
-
