+++ assets/less/sections/sec_login.less
@@ -36,7 +36,6 @@
   grid-template-columns: repeat(4, 1fr);
   justify-items: center;
   user-select: none;
-  
 }
 
 .login__button {
@@ -44,15 +43,15 @@
   font-size: 4rem;
   font-weight: 500;
   text-align: center;
-  height: min-content;    
+  height: min-content;
 }
 
 .login__button_link {
   text-decoration: none;
-  transition: color 0.5s ease-in-out ;
+  transition: color 0.5s ease-in-out;
   color: #ffffff;
   &:hover {
-    color: #f9f2b5
+    color: #f9f2b5;
   }
 }
 
@@ -63,25 +62,25 @@
 
   &_vkontakte {
     &:before {
-      content: '\f189'; //@fa-vk 
+      content: "\f189"; //@fa-vk
     }
   }
 
   &_yandex {
     &:before {
-      content: '\f413'; //@fa-var-yandex
+      content: "\f413"; //@fa-var-yandex
     }
   }
 
   &_google {
     &:before {
-      content: '\f1a0'; //@fa-var-google
+      content: "\f1a0"; //@fa-var-google
     }
   }
 
   &_instagram {
     &:before {
-      content: '\f16d'; // @fa-var-instagram
+      content: "\f16d"; // @fa-var-instagram
     }
   }
 
@@ -89,7 +88,7 @@
     &:before {
       content: "@";
       position: relative;
-      top: -5px
+      top: -5px;
     }
   }
 }
