+++ templates/layout.html.twig
@@ -28,8 +28,7 @@
         crossorigin="anonymous" referrerpolicy="no-referrer" />
      <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css" integrity="sha512-1ycn6IcaQQ40/MKBW2W4Rhis/DbILU74C1vSrLJxCq57o941Ym01SwNsOMqvEBFlcgUa6xLiPY/NS5R+E6ztJQ==" crossorigin="anonymous" referrerpolicy="no-referrer" />
     <!-- Styles ================================================================================== -->
-    <link rel="stylesheet" href="{{ asset('styles.css') }}"/><!-- Main styles from .less -->
-    <link rel="stylesheet" href="{{ asset('main.css') }}"/>
+    <link rel="stylesheet" href="{{ asset('styles.css') }}"/>
     <script src="https://browser.sentry-cdn.com/5.5.0/bundle.min.js" crossorigin="anonymous"></script>
     <script>
       Sentry.init({ dsn: 'https://eb090484615d4e3189a2fae889681e07@sentry.automagistre.ru/5' });
